requirements.txt
google-generativeai>=0.8.0
supabase>=2.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
langchain-text-splitters>=0.2.0
